backend/requirements.txt
alembic==1.17.0
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asyncpg==0.30.0
authx==1.4.3
bcrypt==5.0.0
cffi==2.0.0
click==8.3.0
cryptography==46.0.3
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.120.0
greenlet==3.2.4
h11==0.16.0
idna==3.11
itsdangerous==2.2.0
mako==1.3.10
markupsafe==3.0.3
passlib==1.7.4
psycopg==3.2.12
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.3
pydantic-core==2.41.4
pydantic-settings==2.11.0
pyjwt==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.1
python-jose==3.5.0
pytz==2025.2
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.44
starlette==0.48.0
typing-extensions==4.15.0
typing-inspection==0.4.2
uvicorn==0.38.0
